requirements.txt
pandas
pillow
plotly
google-generativeai
streamlit